requirements.txt
aiofile==3.9.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==26.1.0
Authlib==1.6.9
azure-ai-agents==1.2.0b6
azure-ai-projects==2.0.1
azure-common==1.1.28
azure-core==1.39.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.25.3
azure-monitor-opentelemetry==1.8.7
azure-monitor-opentelemetry-exporter==1.0.0b49
azure-search-documents==11.6.0
azure-storage-blob==12.28.0
backports.tarfile==1.2.0
beartype==0.22.9
blinker==1.9.0
cachetools==7.0.5
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
cyclopts==4.10.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
fastmcp==3.1.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
Jinja2==3.1.6
jiter==0.13.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring==25.7.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
more-itertools==10.8.0
msal==1.35.1
msal-extensions==1.3.1
msrest==0.7.1
narwhals==2.18.0
numpy==2.4.3
oauthlib==3.3.1
ollama==0.6.1
openai==2.29.0
openapi-pydantic==0.5.1
opentelemetry-api==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-asgi==0.61b0
opentelemetry-instrumentation-dbapi==0.61b0
opentelemetry-instrumentation-django==0.61b0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-instrumentation-flask==0.61b0
opentelemetry-instrumentation-logging==0.61b0
opentelemetry-instrumentation-psycopg2==0.61b0
opentelemetry-instrumentation-requests==0.61b0
opentelemetry-instrumentation-urllib==0.61b0
opentelemetry-instrumentation-urllib3==0.61b0
opentelemetry-instrumentation-wsgi==0.61b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
packaging==26.0
pandas==2.3.3
pathable==0.5.0
pillow==12.1.1
platformdirs==4.9.4
protobuf==6.33.6
psutil==7.2.2
py-key-value-aio==0.4.4
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.12.1
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sse-starlette==3.3.3
starlette==1.0.0
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uncalled-for==0.2.0
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
wrapt==1.17.3
zipp==3.23.0
